fix :: 블로그 작성 뷰 이미지 추가 로직 작성
