+++ app/build.gradle.kts
@@ -52,6 +52,6 @@ dependencies {
     implementation (libs.retrofit)
     implementation(libs.converter.gson)
 
-    implementation ("com.github.bumptech.glide:glide:4.9.0")
-    annotationProcessor ("com.github.bumptech.glide:compiler:4.9.0")
+    implementation ("com.github.bumptech.glide:glide:4.14.2")
+    annotationProcessor ("com.github.bumptech.glide:compiler:4.14.2")
 }
