+++ app/src/main/java/com/example/demu_android/feature/Blog/WriteBlogViewModel.kt
@@ -1,5 +1,6 @@
 package com.example.demu_android.feature.Blog
 
+import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
@@ -7,10 +8,12 @@ import com.example.demu_android.feature.recycler.home.WriteBlog.data.Blog
 
 class WriteBlogViewModel : ViewModel() {
     private val _blogList: MutableLiveData<List<Blog>> = MutableLiveData()
-    val blogList: LiveData<List<Blog>> =_blogList
+    var blogList: LiveData<List<Blog>> =_blogList
 
     fun addBlogList(blog : List<Blog>) {
+        Log.d("TEST4",blog.toString())
         _blogList.value = blog
-        _blogList.value = _blogList.value
+        Log.d("TEST2",_blogList.value.toString())
+        Log.d("TEST3",blogList.toString())
     }
 }
