+++ app/src/main/java/com/example/demu_android/feature/recycler/home/WriteBlog/WriteBlogAdaptor.kt
@@ -1,5 +1,6 @@
 package com.example.demu_android.feature.recycler.home.WriteBlog
 
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
